refactor: rename MapUi to MapListSwitch
